tests reorganized with new test class created RefreshControllerRefreshJobIntegrationTest
